requirements.txt
asgiref==3.8.1
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==4.1.13
djongo==1.3.6
dnspython==2.7.0
idna==3.10
lxml==5.3.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pymongo==4.10.1
pytesseract==0.3.13
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.2.4
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.3
